requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

alembic>=0.7.2
greenlet>=0.3.2
keystonemiddleware>=1.0.0
oslo.config>=1.6.0 # Apache-2.0
pbr>=0.6,!=0.7,<1.0
pecan>=0.8.0
stevedore>=1.1.0 # Apache-2.0
SQLAlchemy>=0.9.7,<=0.9.99
